requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
defusedxml==0.7.1
docopt==0.6.2
et_xmlfile==2.0.0
httpie==3.2.4
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.4.3
numpy==2.2.5
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pip-review==1.3.0
Pygments==2.19.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
setuptools==80.7.1
six==1.17.0
tzdata==2025.2
urllib3==2.4.0
veracode-api-signing==25.1.0
